Changes to file structure for the sake of a cleaner looking setup.
Addition of the OFL font license. Miscellaneous changes to bring the font information in line with this license.

Changes of font names. All fonts are now BiangFont, in line the with the Git name and for the sake of clear differentiation.

Edited font files to better match percieved weights. The two base files are now 300 Light and 800 Extra-Bold, rather than Regular and Bold.